requirements.txt
asgiref==3.3.4
Django==3.2.4
django-rest-framework==0.1.0
djangorestframework==3.12.4
factory-boy==3.2.0
Faker==8.8.2
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
